refactored routes, moved css
moved css into its own folder in src/css

refactored routes so they're all in App

made new Home page to replace old App

added back button to the compare page
