+++ src/components/NavBar.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
-import logo from '../images/weight-lifting-logo.png';
-import { BrowserRouter as Router, Route, NavLink } from 'react-router-dom';
 
+import logo from '../images/weight-lifting-logo.png';
 import actionTypes from '../actionTypes';
 import SignOut from './auth/SignOut';
 
@@ -24,40 +23,38 @@ class NavBar extends Component {
 
   render() {
     return (
-      <Router>
-        <nav className="navbar navbar-inverse">
-          <div className="container-fluid">
-            <div className="navbar-header">
-              <a className="navbar-brand" href="#">
-                <img src={logo} alt="logo" style={this.style} />
-              </a>
-            </div>
-            <ul className="nav navbar-nav">
-              <li>
-                <a href="/">Home</a>
-              </li>
-              <li>
-                <a href="/Today" className="fortheloveofgod">
-                  Todays Workout
-                </a>
-              </li>
-              <li>
-                <a href="/Progress">MY GAINZ</a>
-              </li>
-              <li>
-                <a href="/Leaderboards">Leaderboards</a>
-              </li>
-              <li>
-                <a href="/Forums">Forums</a>
-              </li>
-              <li>
-                <a href="/Dashboard">Dashboard</a>
-              </li>
-              {this.renderUserStatus()}
-            </ul>
+      <nav className="navbar navbar-inverse">
+        <div className="container-fluid">
+          <div className="navbar-header">
+            <a className="navbar-brand" href="#">
+              <img src={logo} alt="logo" style={this.style} />
+            </a>
           </div>
-        </nav>
-      </Router>
+          <ul className="nav navbar-nav">
+            <li>
+              <a href="/">Home</a>
+            </li>
+            <li>
+              <a href="/Today" className="fortheloveofgod">
+                Todays Workout
+              </a>
+            </li>
+            <li>
+              <a href="/Progress">MY GAINZ</a>
+            </li>
+            <li>
+              <a href="/Leaderboards">Leaderboards</a>
+            </li>
+            <li>
+              <a href="/Forums">Forums</a>
+            </li>
+            <li>
+              <a href="/Dashboard">Dashboard</a>
+            </li>
+            {this.renderUserStatus()}
+          </ul>
+        </div>
+      </nav>
     );
   }
 }
