+++ src/containers/SetProfile.js
@@ -35,9 +35,8 @@ class SetProfile extends Component {
   handleSubmit = event => {
     event.preventDefault();
 
-    // const userId = 'kG9ajUxIzAPZjF80KKZQ0lYne7t1';
     if (!this.state.weight || !this.state.oneRepMax || !this.state.fullName) {
-      console.log('NOT FILLED OUT YO');
+      alert('Fill out all your stats bro');
     } else {
       const thisUser = firebase.auth().currentUser;
       if (thisUser != null) {
